Refactor expense-related components and API for improved clarity and functionality
- Updated ExpenseNameCell.vue to correct type imports and simplify badge color handling.
- Refactored dashboard.vue to enhance component imports and streamline expense calculations.
- Consolidated expense schemas in types/expense.ts for better organization and type safety.
- Cleaned up expenses store by removing unnecessary budget metrics and optimizing computed properties.
