+++ pages/dashboard.vue
@@ -1,13 +1,14 @@
 <script setup lang="ts">
-import { useRouter } from 'vue-router';
-import { Card, CardHeader, CardTitle, CardContent } from '@/components/ui/card';
-import TableCard from '@/components/TableCard.vue';
 import DateSelector from '@/components/DateSelector.vue';
+import TableCard from '@/components/TableCard.vue';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+import { breakpointsTailwind, useBreakpoints } from '@vueuse/core';
 import { PlusIcon } from 'lucide-vue-next';
-import type { ColumnDefinition, Expense } from '~/types';
+import { useRouter } from 'vue-router';
 import ExpenseNameCell from '~/components/ExpenseNameCell.vue';
-import { useBreakpoints, breakpointsTailwind } from '@vueuse/core';
 import NewExpenseForm from '~/components/newExpenseForm.vue';
+import type { ColumnDefinition } from '~/types';
+import type { Expense } from '~/types/expense';
 
 const router = useRouter();
 const store = useMyExpensesStore();
@@ -29,13 +30,10 @@ watch(isMobile, (value) => {
 
 const getSelectedBudget = computed(() => store.getSelectedBudget);
 const getRemainingDailyBudget = computed(() => store.getRemainingDailyBudget);
-const getRemainingMonthlyBudget = computed(() => store.getRemainingMonthlyBudget);
-const getRemainingBudget = computed(() => store.getRemainingBudget);
-
 if (!getSelectedBudget.value) {
   router.push('/myBudgets');
 }
-const expenses = computed(() => getSelectedBudget?.value?.id ? getExpensesByBudgetId(getSelectedBudget?.value?.id) as Expense[] : []);
+const expenses = computed(() => getSelectedBudget?.value?.id ? getExpensesByBudgetId(getSelectedBudget?.value?.id) : []);
 
 const selectedDate = ref(new Date());
 
@@ -87,28 +85,20 @@ useUpdateMenuElements([
 </script>
 
 <template>
-    <div class="h-full flex flex-col pt-4 gap-6">
-      <h1 class="text-2xl font-bold">{{ getSelectedBudget?.name }} Dashboard</h1>
+  <div class="h-full flex flex-col pt-4 gap-6">
+    <h1 class="text-2xl font-bold">{{ getSelectedBudget?.name }} Dashboard</h1>
 
-      <ResizablePanelGroup 
-      id="handle-demo-group-1" 
-      direction="horizontal" 
-      class="h-full flex !flex-col md:!flex-row gap-4"
-    >
-      <ResizablePanel 
-        id="handle-demo-panel-1" 
-        :default-size="60" 
-        class="flex flex-col gap-4 !basis-auto md:!basis-0"
-      >
+    <ResizablePanelGroup id="handle-demo-group-1" direction="horizontal"
+      class="h-full flex !flex-col md:!flex-row gap-4">
+      <ResizablePanel id="handle-demo-panel-1" :default-size="60" class="flex flex-col gap-4 !basis-auto md:!basis-0">
         <DateSelector v-model:selectedDate="selectedDate" />
         <div class="flex flex-row gap-4 w-full">
           <Card class="w-full">
             <CardHeader>
               <CardTitle>Total Expenses</CardTitle>
             </CardHeader>
-            <CardContent
-              :class="{ 'text-destructive-foreground': getRemainingBudget < 0, 'text-success': getRemainingBudget >= 0 }">
-              {{ getRemainingBudget }}
+            <CardContent>
+              0
             </CardContent>
           </Card>
 
@@ -116,9 +106,8 @@ useUpdateMenuElements([
             <CardHeader>
               <CardTitle>Monthly Budget</CardTitle>
             </CardHeader>
-            <CardContent
-              :class="{ 'text-destructive-foreground': getRemainingMonthlyBudget < 0, 'text-success': getRemainingMonthlyBudget >= 0 }">
-              {{ getRemainingMonthlyBudget }}
+            <CardContent>
+              0
             </CardContent>
           </Card>
 
@@ -128,7 +117,7 @@ useUpdateMenuElements([
             </CardHeader>
             <CardContent
               :class="{ 'text-destructive-foreground': getRemainingDailyBudget < 0, 'text-success': getRemainingDailyBudget >= 0 }">
-              {{ getRemainingDailyBudget }}
+              $ {{ getRemainingDailyBudget.toFixed(2) }}
             </CardContent>
           </Card>
         </div>
@@ -140,11 +129,7 @@ useUpdateMenuElements([
         </Card>
       </ResizablePanel>
       <ResizableHandle id="handle-demo-handle-1" class="hidden md:flex" />
-      <ResizablePanel 
-        id="handle-demo-panel-2" 
-        :default-size="40"
-        class="!basis-auto md:!basis-0"
-      >
+      <ResizablePanel id="handle-demo-panel-2" :default-size="40" class="!basis-auto md:!basis-0">
         <TableCard title="Expenses" :content="expenses || []" :config="expensesConfig"
           v-model:condensed-mode="condensedMode" class="h-[54dvh] md:h-[85dvh]">
           <div class="flex flex-col items-center w-full">
@@ -158,5 +143,5 @@ useUpdateMenuElements([
 
     <!-- New Expense Form Dialog -->
     <NewExpenseForm v-model="showExpenseForm" />
-    </div>
+  </div>
 </template>
